Scripts to test SMP/E PAX
The script to run is `install-SMPE-PAX.sh` which reads the README for the GUNZIP job, runs that and then tailors and runs the resulting SMP/E JCL jobs.

To re-run, execute the `delete-relfiles.sh` script, which deletes the output of the SMP/E install and creates a directory for the next time.  The creation of the directory should be moved to occur before running `install-SMPE-PAX.sh`.

This is a work-in-progress.